requirements.txt
django-tables2==1.1.1
django-filter==0.12.0
django-bootstrap3==6.2.2